Added Computer Area Map. Also removed default hotkeys Berserk, Antirad.
Now only Nightvision has a hotkey of 'N' that is automatically defined
if still free. The others are cheats really (a philosophical point yes,
but I believe a space marine has light amp on him and does not need
to die 20 times until he knows monster spawn points be heart)
and should not be automatically active.